dt-bindings: arm: Add scmi_devid property description for SCMI
Document scmi_devid property for the devices, using SCMI protocol
to work with clocks/resets/power-domains etc. This property is intended
to set the device id, which should be used to manage device permissions
in the firmware. Device permissions management is descibed in DEN 0056,
Section 4.2.2.10 [0].

This property is useful for the virtualized systems, when several agents
are running on the same platform. Agent term is descibed in Section
4.1.1 [0].

[0] https://developer.arm.com/documentation/den0056/latest

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>