opp: Add support for 'advanced' Energy Model in DT
The Energy Model (EM) can be created based on DT entry:
'dynamic-power-coefficient'. It's a 'simple' EM which is limited to the
dynamic power. It has to fit into the math formula which requires also
information about voltage. Some of the platforms don't expose voltage
information, thus it's not possible to use EM registration using DT.

This patch aims to fix it. It introduces new implementation of the EM
registration callback. The new mechanism parses EM array specified in DT
which contains a set of tuples: frequency (in kHz) and power (uW).
It also allows to register 'advanced' EM, which models total power
(static + dynamic), so better reflects real HW.

Signed-off-by: Lukasz Luba <lukasz.luba@arm.com>