VFS: support parallel updates in the one directory.
Hi Al,
 I wonder if you might find time to have a look at this patch.  It
 allows concurrent updates to a single directory.  This can result in
 substantial throughput improvements when the application uses multiple
 threads to create lots of files in the one directory, and there is
 noticeable per-create latency, as there can be with NFS to a remote
 server.
Thanks,
NeilBrown

Some filesystems can support parallel modifications to a directory,
either because the modification happen on a remote server which does its
own locking (e.g.  NFS) or because they can internally lock just a part
of a directory (e.g.  many local filesystems, with a bit of work - the
lustre project has patches for ext4 to support concurrent updates).

To allow this, we introduce VFS support for parallel modification:
unlink (including rmdir) and create.  Parallel rename is not (yet)
supported.

If a filesystem supports parallel modification in a given directory, it
sets S_PAR_UNLINK on the inode for that directory.  lookup_open() and
the new lookup_hash_modify() (similar to __lookup_hash()) notice the
flag and take a shared lock on the directory, and rely on a lock-bit in
d_flags, much like parallel lookup relies on DCACHE_PAR_LOOKUP.

The new DCACHE_PAR_UPDATE is most significant bit in d_flags, so we have
nearly run out of flags.  0x08000000 is still unused .... for now.

Once a dentry for the target name has been obtained, DCACHE_PAR_UPDATE
is set on it, waiting if necessary if it is already set.  Once this is
set, the thread has exclusive access to the name and can call into the
filesystem to perform the required action.

We use wait_var_event() to wait for DCACHE_PAR_UPDATE to be cleared rather
than trying to overload the wq used for DCACHE_PAR_LOOKUP.

Some filesystems do *not* complete the lookup that precedes a create,
but leave the dentry d_in_lookup() and unhashed, so often a dentry will
have both DCACHE_PAR_LOOKUP and DCACHE_PAR_UPDATE set at the same time.
To allow for this, we need the 'wq' that is passed to d_alloc_parallel()
(and used when DCACHE_PAR_LOOKUP is cleared), to continue to exist until
the creation is complete, which may be after filename_create()
completes.  So a wq now needs to be passed to filename_create() when
parallel modification is attempted.

Waiting for DCACHE_PAR_UPDATE can sometimes result in the dentry
changing - either through unlink or rename.  This requires that the
lookup be retried.  This in turn requires that the wq be reused.
The use for DECLARE_WAITQUEUE() in d_wait_lookup() and the omission of
remove_wait_queue() means that the after the wake_up_all() in
__d_lookup_down(), the wait_queue_head list will be undefined.  So we change
d_wait_lookup() to use DEFINE_WAIT(), so that autoremove_wake_function()
is used for wakeups, and the wait_queue_head remains well defined.

NFS can easily support parallel updates as the locking is done on the
server, so this patch enable parallel updates for NFS.  This would be a
separate patch in the final submission.

NFS unlink needs to block concurrent opens() once it decides to actually
unlink the file, rather the rename it to .nfsXXXX (aka sillyrename).
It currently does this by temporarily unhashing the dentry and relying
on the exclusive lock on the directory to block a ->lookup().  That
doesn't work now that unlink uses a shared lock, so an alternate
approach is needed.

__nfs_lookup_revalidate (->d_revalidate) now blocks if DCACHE_PAR_UPDATE
is set, and if nfs_unlink() happens to be called with an exclusive lock
and DCACHE_PAR_UPDATE is not it, it sets during the potential race window.

I'd rather use some other indicator in the dentry to tell
_nfs_lookup_revalidate() to wait, but we are nearly out of d_flags bits,
and NFS does have a general-purpose d_fsdata.

NFS "silly-rename" may now be called with only a shared lock on the
directory, so it needs a bit of extra care to get exclusive access to
the new name. d_lock_update_nested() and d_unlock_update() help here.

Signed-off-by: NeilBrown <neilb@suse.com>