MAINTAINERS, SLAB: Add Roman as reviewer, git tree
The slab code has an overlap with kmem accounting, where Roman has done
a lot of work recently and it would be useful to make sure he's CC'd on
patches that potentially affect it. Thus add him as a reviewer for the
SLAB subsystem.

Also while at it, add the link to slab git tree.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>