scsi: remove the cmd field from struct scsi_request
Now that each scsi_request is backed by a scsi_cmnd, there is no
need indirect the CDB storage.  Change all submitters of SCSI
passthrough request to store the CDB information directly in the
scsi_cmnd, and while doing so allocate the full 32 bytes that cover
all Linux supported SCSI hosts instead of requiring dynamic allocation
for > 16 byte CDBs.  On 64-bit systems this does not change the size of
the scsi_cmnd at all, while on 32-bit systems it slightly increases it
for now, but that increase will be made up by the removal of the
remaining scsi_request fields.

Signed-off-by: Christoph Hellwig <hch@lst.de>