PCI: Apply quirk_amd_harvest_no_ats to all navi10 and 14 asics
There are enough vbios escapes without the proper workaround
that some users still hit this.  MS never productized ATS on
windows so OEM platforms that were windows only didn't always
validate ATS.

The advantages of ATS are not worth it compared to the potential
instabilities on harvested boards.  Just disable ATS on all navi10
and 14 boards.

Bug: https://gitlab.freedesktop.org/drm/amd/-/issues/1760
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>