Docs/vm/damon: Call low level monitoring primitives the operations
DAMON code calls the low level monitoring primitives implementations the
monitoring operations.  The documentation would have no problem at still
calling those primitives implementation because there is no real
difference in the concepts, but making it more consistent with the code
would make it better.  This commit therefore convert sentences in the
doc specifically pointing the implementations of the primitives to call
it monitoring operations.

Signed-off-by: SeongJae Park <sj@kernel.org>