locking/lockdep: Optimize look_up_lock_class()
The lockdep_map structure has a built-in cache for the first two
subclasses.  However, when lockdep_init_map_type() is called with a
subclass other than 0, it forces class_cache[0] to hold class for a
non-0 subclass. That messes up the utility of the cache as you never
know if class_cache[0] holds a class for subclass 0.

Use the 1-byte hole in the lockdep_map structure to hold the
subclass information in class_cache[0] so that we can reliably use the
class_cache[] information to get the class based on the given subclass
quicker in look_up_lock_class().

Signed-off-by: Waiman Long <longman@redhat.com>